Show historical logs for notifications
